app/requirements.txt
scikit-learn
pandas
joblib
# prometheus-fastapi-instrumentator
uvicorn
fastapi
python-dotenv